pyproject.toml
[tool.poetry]
name = "webtronics_task"
version = "0.1.0"
description = ""
authors = ["Alexey Bystrov <gipard446@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
SQLAlchemy = "^2.0.18"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
redis = "^4.6.0"
aioredis = "^2.0.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
